fix ops confirmation
