+++ internal/cli/cmd/accounts/delete.go
@@ -20,11 +20,11 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package accounts
 
 import (
+	"github.com/apecloud/kubeblocks/internal/cli/util/prompt"
 	"github.com/spf13/cobra"
 	"k8s.io/cli-runtime/pkg/genericclioptions"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 
-	"github.com/apecloud/kubeblocks/internal/cli/delete"
 	channelutil "github.com/apecloud/kubeblocks/internal/sqlchannel/util"
 )
 
@@ -56,7 +56,7 @@ func (o *DeleteUserOptions) Validate(args []string) error {
 	if o.AutoApprove {
 		return nil
 	}
-	if err := delete.Confirm([]string{o.info.UserName}, o.In, false); err != nil {
+	if err := prompt.Confirm([]string{o.info.UserName}, o.In, false); err != nil {
 		return err
 	}
 	return nil
