+++ internal/cli/delete/delete.go
@@ -21,12 +21,8 @@ package delete
 
 import (
 	"fmt"
-	"io"
-	"sort"
-	"strings"
 
 	"github.com/spf13/cobra"
-	"golang.org/x/exp/slices"
 	metav1 "k8s.io/apimachinery/pkg/apis/meta/v1"
 	"k8s.io/apimachinery/pkg/runtime"
 	"k8s.io/apimachinery/pkg/runtime/schema"
@@ -34,7 +30,6 @@ import (
 	"k8s.io/cli-runtime/pkg/resource"
 	cmdutil "k8s.io/kubectl/pkg/cmd/util"
 
-	"github.com/apecloud/kubeblocks/internal/cli/printer"
 	"github.com/apecloud/kubeblocks/internal/cli/util"
 	"github.com/apecloud/kubeblocks/internal/cli/util/prompt"
 )
@@ -157,7 +152,7 @@ func (o *DeleteOptions) complete() error {
 		if len(names) == 0 {
 			names = o.Names
 		}
-		if err = Confirm(names, o.In, len(o.LabelSelector) != 0); err != nil {
+		if err = prompt.Confirm(names, o.In, len(o.LabelSelector) != 0); err != nil {
 			return err
 		}
 	}
@@ -239,29 +234,4 @@ func (o *DeleteOptions) postDeleteResource(object runtime.Object) error {
 		return o.PostDeleteHook(o, object)
 	}
 	return nil
-}
-
-// Confirm let user double-check for the cluster ops
-// if isLabelSelector is true, it will output the details information for confirmation
-func Confirm(names []string, in io.Reader, isLabelSelector bool) error {
-	if len(names) == 0 {
-		return nil
-	}
-	// if the resources were selected by label, user don't know their names we should give a prompt
-	// and only delete for cluster may use it
-	if isLabelSelector {
-		fmt.Printf("These clusters will be deleted:[%s]\n", printer.BoldRed(strings.Join(names, " ")))
-	}
-	// '\n' in NewPrompt will break the output
-	_, err := prompt.NewPrompt("Please type the name again(separate with white space when more than one):",
-		func(entered string) error {
-			enteredNames := strings.Split(entered, " ")
-			sort.Strings(names)
-			sort.Strings(enteredNames)
-			if !slices.Equal(names, enteredNames) {
-				return fmt.Errorf("typed \"%s\" does not match \"%s\"", entered, strings.Join(names, " "))
-			}
-			return nil
-		}, in).Run()
-	return err
 }
