+++ internal/cli/util/prompt/prompt.go
@@ -20,9 +20,15 @@ along with this program.  If not, see <http://www.gnu.org/licenses/>.
 package prompt
 
 import (
+	"fmt"
 	"io"
+	"sort"
+	"strings"
 
 	"github.com/manifoldco/promptui"
+	"golang.org/x/exp/slices"
+
+	"github.com/apecloud/kubeblocks/internal/cli/printer"
 )
 
 func NewPrompt(label string, validate promptui.ValidateFunc, in io.Reader) *promptui.Prompt {
@@ -48,4 +54,29 @@ func NewPrompt(label string, validate promptui.ValidateFunc, in io.Reader) *prom
 		Validate:  validate,
 	}
 	return &p
+}
+
+// Confirm let user double-check for the cluster ops
+// if isLabelSelector is true, it will output the details information for confirmation
+func Confirm(names []string, in io.Reader, isLabelSelector bool) error {
+	if len(names) == 0 {
+		return nil
+	}
+	// if the resources were selected by label, user don't know their names we should give a prompt
+	// and only delete for cluster may use it
+	if isLabelSelector {
+		fmt.Printf("These clusters will be deleted:[%s]\n", printer.BoldRed(strings.Join(names, " ")))
+	}
+	// '\n' in NewPrompt will break the output
+	_, err := NewPrompt("Please type the name again(separate with white space when more than one):",
+		func(entered string) error {
+			enteredNames := strings.Split(entered, " ")
+			sort.Strings(names)
+			sort.Strings(enteredNames)
+			if !slices.Equal(names, enteredNames) {
+				return fmt.Errorf("typed \"%s\" does not match \"%s\"", entered, strings.Join(names, " "))
+			}
+			return nil
+		}, in).Run()
+	return err
 }
